+++ src/main/java/com/vnaskos/pdfguru/execution/FileNamer.java
@@ -1,4 +1,4 @@
-package com.vnaskos.pdfguru.execute;
+package com.vnaskos.pdfguru.execution;
 
 import java.io.File;
