+++ src/main/java/com/vnaskos/pdfguru/execution/Finishable.java
@@ -1,4 +1,4 @@
-package com.vnaskos.pdfguru.execute;
+package com.vnaskos.pdfguru.execution;
 
 public interface Finishable {
