+++ src/main/java/com/vnaskos/pdfguru/execution/PagePatternTranslator.java
@@ -1,4 +1,4 @@
-package com.vnaskos.pdfguru.execute;
+package com.vnaskos.pdfguru.execution;
 
 import java.util.ArrayList;
 import java.util.List;
