+++ src/test/java/com/vnaskos/pdfguru/execution/ProcessOrchestratorTest.java
@@ -1,4 +1,4 @@
-package com.vnaskos.pdfguru.execute;
+package com.vnaskos.pdfguru.execution;
 
 import com.vnaskos.pdfguru.input.items.InputItem;
 import org.junit.Test;
